+++ src/main/java/hs/kr/dgsw/spring0318/CalculatorService.java
@@ -1,5 +1,5 @@
 package hs.kr.dgsw.spring0318;
 
 public interface CalculatorService {
-    String calculate(int fv, String op, int lv); //첫번째 숫자, 연산자, 두번째 숫자
+    String calculate(double fv, String op, double lv); //첫번째 숫자, 연산자, 두번째 숫자
 }
